Week7_Lab/requirements.txt
groq==1.0.0
snowflake-connector-python==3.18.0
python-dotenv==1.0.0
streamlit==1.54.0
pandas==2.2.3
altair==6.0.0
scikit-learn==1.7.2
numpy==1.26.4
pyyaml==6.0.2
